+++ packages/library-base/src/commonMain/kotlin/io/realm/TypedRealm.kt
@@ -1,5 +1,6 @@
 package io.realm
 
+import io.realm.query.RealmQuery
 import kotlin.reflect.KClass
 
 /**
@@ -8,26 +9,32 @@ import kotlin.reflect.KClass
 interface TypedRealm : BaseRealm {
 
     /**
-     * Returns the results of querying for all objects of a specific type.
+     * Returns a [RealmQuery] matching the predicate represented by [query].
      *
-     * For a [Realm] instance this reflects the state of the realm at the invocation time, thus
-     * the results will not change on updates to the Realm. For a [MutableRealm] the result is live
-     * and will in fact reflect updates to the [MutableRealm].
+     * For a [Realm] instance this reflects the state of the Realm at the invocation time, this
+     * the results obtained from the query will not change on updates to the Realm. For a
+     * [MutableRealm] the query will produce live results and will in fact reflect updates to the
+     * [MutableRealm].
      *
-     * @param clazz The class of the objects to query for.
-     * @return The result of the query.
+     * @param query the Realm Query Language predicate to append.
+     * @param args Realm values for the predicate.
      */
-    open fun <T : RealmObject> objects(clazz: KClass<T>): RealmResults<T>
+    fun <T : RealmObject> query(
+        clazz: KClass<T>,
+        query: String = "TRUEPREDICATE",
+        vararg args: Any?
+    ): RealmQuery<T>
 }
 
 /**
- * Returns the results of querying for all objects of a specific type.
+ * Returns a [RealmQuery] matching the predicate represented by [query].
  *
- * Reified convenience wrapper of [TypedRealm.objects].
+ * Reified convenience wrapper of [TypedRealm.query].
  *
- * @param T Type of the object to query for.
- * @return The result of the query.
+ * @param query the Realm Query Language predicate to append.
+ * @param args Realm values for the predicate.
  */
-inline fun <reified T : RealmObject> TypedRealm.objects(): RealmResults<T> {
-    return this.objects(T::class)
-}
+inline fun <reified T : RealmObject> TypedRealm.query(
+    query: String = "TRUEPREDICATE",
+    vararg args: Any?
+): RealmQuery<T> = query(T::class, query, *args)