+++ packages/library-base/src/commonMain/kotlin/io/realm/internal/BaseRealmImpl.kt
@@ -19,10 +19,11 @@ import io.realm.BaseRealm
 import io.realm.Callback
 import io.realm.Cancellable
 import io.realm.RealmObject
-import io.realm.RealmResults
 import io.realm.internal.interop.NativePointer
 import io.realm.internal.interop.RealmInterop
+import io.realm.internal.query.ObjectQuery
 import io.realm.internal.schema.RealmSchemaImpl
+import io.realm.query.RealmQuery
 import io.realm.schema.RealmSchema
 import kotlinx.coroutines.flow.Flow
 import kotlin.reflect.KClass
@@ -34,7 +35,8 @@ abstract class BaseRealmImpl internal constructor(
 ) : BaseRealm, RealmStateHolder {
 
     private companion object {
-        private const val OBSERVABLE_NOT_SUPPORTED_MESSAGE = "Observing changes are not supported by this Realm."
+        private const val OBSERVABLE_NOT_SUPPORTED_MESSAGE =
+            "Observing changes are not supported by this Realm."
     }
 
     /**
@@ -72,23 +74,14 @@ abstract class BaseRealmImpl internal constructor(
         return RealmSchemaImpl.fromRealm(realmReference.dbPointer)
     }
 
-    open fun <T : RealmObject> objects(clazz: KClass<T>): RealmResults<T> {
-        // Use same reference through out all operations to avoid locking
-        val realmReference = this.realmReference
-        realmReference.checkClosed()
-        return RealmResultsImpl.fromQuery(
-            realmReference,
-            RealmInterop.realm_query_parse(
-                realmReference.dbPointer,
-                clazz.simpleName!!,
-                "TRUEPREDICATE"
-            ),
-            clazz,
-            configuration.mediator
-        )
-    }
+    open fun <T : RealmObject> query(
+        clazz: KClass<T>,
+        query: String,
+        vararg args: Any?
+    ): RealmQuery<T> =
+        ObjectQuery(realmReference, clazz, configuration.mediator, null, query, *args)
 
-    internal open fun <T> registerObserver(t: Observable<T>): Flow<T> {
+    internal open fun <T> registerObserver(t: Thawable<T>): Flow<T> {
         throw NotImplementedError(OBSERVABLE_NOT_SUPPORTED_MESSAGE)
     }