+++ packages/library-base/src/commonMain/kotlin/io/realm/internal/RealmObjectInternal.kt
@@ -37,7 +37,7 @@ import kotlin.reflect.KClass
  * [RealmObject].
  */
 @Suppress("VariableNaming")
-interface RealmObjectInternal : RealmObject, RealmStateHolder, io.realm.internal.interop.RealmObjectInterop, Observable<RealmObjectInternal> {
+interface RealmObjectInternal : RealmObject, RealmStateHolder, io.realm.internal.interop.RealmObjectInterop, Observable<RealmObjectInternal>, Flowable<RealmObjectInternal> {
     // Names must match identifiers in compiler plugin (plugin-compiler/io.realm.compiler.Identifiers.kt)
 
     // Reference to the public Realm instance and internal transaction to which the object belongs.
@@ -78,7 +78,7 @@ interface RealmObjectInternal : RealmObject, RealmStateHolder, io.realm.internal
         }
     }
 
-    override fun thaw(liveRealm: RealmReference): Observable<RealmObjectInternal>? {
+    override fun thaw(liveRealm: RealmReference): RealmObjectInternal? {
         @Suppress("UNCHECKED_CAST")
         val type: KClass<out RealmObject> = this::class
         val mediator = `$realm$Mediator`!!
@@ -113,7 +113,7 @@ interface RealmObjectInternal : RealmObject, RealmStateHolder, io.realm.internal
         }
     }
 
-    override fun observe(): Flow<RealmObjectInternal> {
+    override fun asFlow(): Flow<RealmObjectInternal> {
         return this.`$realm$Owner`!!.owner.registerObserver(this)
     }
 }