+++ test/base/src/androidTest/kotlin/io/realm/test/shared/RealmListTests.kt
@@ -27,7 +27,8 @@ import io.realm.entities.list.Level1
 import io.realm.entities.list.Level2
 import io.realm.entities.list.Level3
 import io.realm.entities.list.RealmListContainer
-import io.realm.objects
+import io.realm.query
+import io.realm.query.find
 import io.realm.realmListOf
 import io.realm.test.platform.PlatformUtils
 import io.realm.test.util.TypeDescriptor
@@ -117,12 +118,15 @@ class RealmListTests {
             copyToRealm(level1_2) // this includes the graph of all 6 objects
         }
 
-        val objectsL1: RealmResults<Level1> =
-            realm.objects<Level1>().query("name BEGINSWITH \"l\" SORT(name ASC)")
-        val objectsL2: RealmResults<Level2> =
-            realm.objects<Level2>().query("name BEGINSWITH \"l\" SORT(name ASC)")
-        val objectsL3: RealmResults<Level3> =
-            realm.objects<Level3>().query("name BEGINSWITH \"l\" SORT(name ASC)")
+        val objectsL1: RealmResults<Level1> = realm.query<Level1>()
+            .query("""name BEGINSWITH "l" SORT(name ASC)""")
+            .find()
+        val objectsL2: RealmResults<Level2> = realm.query<Level2>()
+            .query("""name BEGINSWITH "l" SORT(name ASC)""")
+            .find()
+        val objectsL3: RealmResults<Level3> = realm.query<Level3>()
+            .query("""name BEGINSWITH "l" SORT(name ASC)""")
+            .find()
 
         assertEquals(2, objectsL1.count())
         assertEquals(2, objectsL2.count())
@@ -548,9 +552,12 @@ internal abstract class ManagedListTester<T>(
                     .addAll(dataSet)
             }
 
-            val list = realm.objects<RealmListContainer>()
+            val list = realm.query<RealmListContainer>()
                 .first()
-                .let { typeSafetyManager.getList(it) }
+                .find { listContainer ->
+                    assertNotNull(listContainer)
+                    typeSafetyManager.getList(listContainer)
+                }
 
             realm.close()
 
@@ -811,7 +818,10 @@ internal abstract class ManagedListTester<T>(
     // Retrieves the list again but this time from Realm to check the getter is called correctly
     private fun assertListAndCleanup(assertion: (RealmList<T>) -> Unit) {
         realm.writeBlocking {
-            val container = this.objects<RealmListContainer>().first()
+            val container = this.query<RealmListContainer>()
+                .first()
+                .find()
+            assertNotNull(container)
             val list = typeSafetyManager.getList(container)
 
             // Assert
@@ -825,7 +835,10 @@ internal abstract class ManagedListTester<T>(
     }
 
     private fun assertContainerAndCleanup(assertion: (RealmListContainer) -> Unit) {
-        val container = realm.objects<RealmListContainer>().first()
+        val container = realm.query<RealmListContainer>()
+            .first()
+            .find()
+        assertNotNull(container)
 
         // Assert
         errorCatcher {
@@ -875,19 +888,20 @@ internal val BYTE_VALUES = listOf<Byte>(1, 2)
 internal val FLOAT_VALUES = listOf(1F, 2F)
 internal val DOUBLE_VALUES = listOf(1.0, 2.0)
 internal val BOOLEAN_VALUES = listOf(true, false)
-internal val TIMESTAMP_VALUES = listOf(RealmInstant.fromEpochSeconds(0, 0), RealmInstant.fromEpochSeconds(42, 420))
+internal val TIMESTAMP_VALUES =
+    listOf(RealmInstant.fromEpochSeconds(0, 0), RealmInstant.fromEpochSeconds(42, 420))
 internal val OBJECT_VALUES = listOf(
     RealmListContainer().apply { stringField = "A" },
     RealmListContainer().apply { stringField = "B" }
 )
 
-internal val NULLABLE_CHAR_VALUES = listOf('a', 'b', null)
-internal val NULLABLE_STRING_VALUES = listOf("ABC", "BCD", null)
-internal val NULLABLE_INT_VALUES = listOf(1, 2, null)
-internal val NULLABLE_LONG_VALUES = listOf<Long?>(1, 2, null)
-internal val NULLABLE_SHORT_VALUES = listOf<Short?>(1, 2, null)
-internal val NULLABLE_BYTE_VALUES = listOf<Byte?>(1, 2, null)
-internal val NULLABLE_FLOAT_VALUES = listOf(1F, 2F, null)
-internal val NULLABLE_DOUBLE_VALUES = listOf(1.0, 2.0, null)
-internal val NULLABLE_BOOLEAN_VALUES = listOf(true, false, null)
-internal val NULLABLE_TIMESTAMP_VALUES = listOf(RealmInstant.fromEpochSeconds(0, 0), RealmInstant.fromEpochSeconds(42, 420), null)
+internal val NULLABLE_CHAR_VALUES = CHAR_VALUES + null
+internal val NULLABLE_STRING_VALUES = STRING_VALUES + null
+internal val NULLABLE_INT_VALUES = INT_VALUES + null
+internal val NULLABLE_LONG_VALUES = LONG_VALUES + null
+internal val NULLABLE_SHORT_VALUES = SHORT_VALUES + null
+internal val NULLABLE_BYTE_VALUES = BYTE_VALUES + null
+internal val NULLABLE_FLOAT_VALUES = FLOAT_VALUES + null
+internal val NULLABLE_DOUBLE_VALUES = DOUBLE_VALUES + null
+internal val NULLABLE_BOOLEAN_VALUES = BOOLEAN_VALUES + null
+internal val NULLABLE_TIMESTAMP_VALUES = TIMESTAMP_VALUES + null