+++ test/base/src/androidTest/kotlin/io/realm/test/shared/SampleTests.kt
@@ -23,12 +23,13 @@ package io.realm.test.shared
 import io.realm.Realm
 import io.realm.RealmConfiguration
 import io.realm.RealmInstant
-import io.realm.RealmResults
 import io.realm.delete
 import io.realm.entities.Sample
 import io.realm.internal.RealmObjectCompanion
 import io.realm.internal.platform.realmObjectCompanion
 import io.realm.internal.realmObjectCompanion
+import io.realm.query
+import io.realm.query.find
 import io.realm.test.platform.PlatformUtils
 import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
@@ -46,7 +47,10 @@ class SampleTests {
     @BeforeTest
     fun setup() {
         tmpDir = PlatformUtils.createTempDir()
-        val configuration = RealmConfiguration.Builder(schema = setOf(Sample::class)).path("$tmpDir/default.realm").build()
+        val configuration =
+            RealmConfiguration.Builder(schema = setOf(Sample::class))
+                .path("$tmpDir/default.realm")
+                .build()
         realm = Realm.open(configuration)
     }
 
@@ -124,23 +128,28 @@ class SampleTests {
             copyToRealm(Sample()).run { stringField = "Hello, Realm!" }
         }
 
-        val objects1: RealmResults<Sample> = realm.objects(Sample::class)
-        assertEquals(2, objects1.size)
+        realm.query<Sample>()
+            .find { results ->
+                assertEquals(2, results.size)
+            }
 
-        val objects2: RealmResults<Sample> =
-            realm.objects(Sample::class).query("stringField == $0", s)
-        assertEquals(1, objects2.size)
-        for (sample in objects2) {
-            assertEquals(s, sample.stringField)
-        }
+        realm.query<Sample>("stringField == $0", s)
+            .find { results ->
+                assertEquals(1, results.size)
+                for (sample in results) {
+                    assertEquals(s, sample.stringField)
+                }
+            }
     }
 
     @Test
     fun query_parseErrorThrows() {
-        val objects: RealmResults<Sample> = realm.objects(Sample::class)
-        assertFailsWith<IllegalArgumentException> {
-            objects.query("name == str")
-        }
+        realm.query<Sample>()
+            .find { results ->
+                assertFailsWith<IllegalArgumentException> {
+                    results.query("name == str")
+                }
+            }
     }
 
     @Test
@@ -150,17 +159,22 @@ class SampleTests {
             copyToRealm(Sample()).run { stringField = "Hello, Realm!" }
         }
 
-        val objects1: RealmResults<Sample> = realm.objects(Sample::class)
-        assertEquals(2, objects1.size)
+        realm.query<Sample>()
+            .find { results ->
+                assertEquals(2, results.size)
+            }
 
         realm.writeBlocking {
-            objects(Sample::class).delete()
+            query<Sample>()
+                .find()
+                .delete()
         }
 
-        assertEquals(0, realm.objects(Sample::class).size)
+        assertEquals(0, realm.query<Sample>().find().size)
     }
 
     @Test
+    @Suppress("LongMethod")
     fun primitiveTypes() {
         realm.writeBlocking {
             copyToRealm(Sample()).apply {
@@ -177,49 +191,71 @@ class SampleTests {
             }
         }
 
-        var objects: RealmResults<Sample> = realm.objects(Sample::class)
-        assertEquals(1, objects.size)
+        realm.query<Sample>()
+            .find { objects ->
+                assertEquals(1, objects.size)
 
-        assertEquals("Realm Kotlin", objects[0].stringField)
-        assertEquals(0xb, objects[0].byteField)
-        assertEquals('b', objects[0].charField)
-        assertEquals(1, objects[0].shortField)
-        assertEquals(2, objects[0].intField)
-        assertEquals(1024, objects[0].longField)
-        assertFalse(objects[0].booleanField)
-        assertEquals(1.99f, objects[0].floatField)
-        assertEquals(1.19851106, objects[0].doubleField)
-        assertEquals(RealmInstant.fromEpochSeconds(42, 420), objects[0].timestampField)
+                assertEquals("Realm Kotlin", objects[0].stringField)
+                assertEquals(0xb, objects[0].byteField)
+                assertEquals('b', objects[0].charField)
+                assertEquals(1, objects[0].shortField)
+                assertEquals(2, objects[0].intField)
+                assertEquals(1024, objects[0].longField)
+                assertFalse(objects[0].booleanField)
+                assertEquals(1.99f, objects[0].floatField)
+                assertEquals(1.19851106, objects[0].doubleField)
+                assertEquals(RealmInstant.fromEpochSeconds(42, 420), objects[0].timestampField)
+            }
 
         // querying on each type
-        objects = realm.objects(Sample::class).query("stringField == $0", "Realm Kotlin") // string
-        assertEquals(1, objects.size)
+        realm.query<Sample>("stringField == $0", "Realm Kotlin") // string
+            .find { objects ->
+                assertEquals(1, objects.size)
+            }
 
-        objects = realm.objects(Sample::class).query("byteField == $0", 0xb) // byte
-        assertEquals(1, objects.size)
+        realm.query<Sample>("byteField == $0", 0xb) // byte
+            .find { objects ->
+                assertEquals(1, objects.size)
+            }
 
-        objects = realm.objects(Sample::class).query("charField == $0", 'b') // char
-        assertEquals(1, objects.size)
+        realm.query<Sample>("charField == $0", 'b') // char
+            .find { objects ->
+                assertEquals(1, objects.size)
+            }
 
-        objects = realm.objects(Sample::class).query("shortField == $0", 1) // short
-        assertEquals(1, objects.size)
+        realm.query<Sample>("shortField == $0", 1) // short
+            .find { objects ->
+                assertEquals(1, objects.size)
+            }
 
-        objects = realm.objects(Sample::class).query("intField == $0", 2) // int
-        assertEquals(1, objects.size)
+        realm.query<Sample>("intField == $0", 2) // int
+            .find { objects ->
+                assertEquals(1, objects.size)
+            }
 
-        objects = realm.objects(Sample::class).query("longField == $0", 1024) // long
-        assertEquals(1, objects.size)
+        realm.query<Sample>("longField == $0", 1024) // long
+            .find { objects ->
+                assertEquals(1, objects.size)
+            }
 
-        objects = realm.objects(Sample::class).query("booleanField == false") // FIXME query("booleanField == $0", false) is not working
-        assertEquals(1, objects.size)
+        realm.query<Sample>("booleanField == false") // FIXME query("booleanField == $0", false) is not working
+            .find { objects ->
+                assertEquals(1, objects.size)
+            }
 
-        objects = realm.objects(Sample::class).query("floatField == $0", 1.99f)
-        assertEquals(1, objects.size)
+        realm.query<Sample>("floatField == $0", 1.99f)
+            .find { objects ->
+                assertEquals(1, objects.size)
+            }
 
-        objects = realm.objects(Sample::class).query("doubleField == $0", 1.19851106)
-        assertEquals(1, objects.size)
+        realm.query<Sample>("doubleField == $0", 1.19851106)
+            .find { objects ->
+                assertEquals(1, objects.size)
+            }
 
-        objects = realm.objects(Sample::class).query("timestampField == $0", RealmInstant.fromEpochSeconds(42, 420))
-        assertEquals(1, objects.size)
+        realm.query<Sample>("timestampField == $0", RealmInstant.fromEpochSeconds(42, 420))
+            .find { objects ->
+                assertEquals(1, objects.size)
+            }
     }
 }