+++ test/base/src/macosTest/kotlin/io/realm/test/MemoryTests.kt
@@ -19,6 +19,7 @@ package io.realm.test
 import io.realm.Realm
 import io.realm.RealmConfiguration
 import io.realm.entities.Sample
+import io.realm.query
 import io.realm.test.platform.PlatformUtils.createTempDir
 import io.realm.test.platform.PlatformUtils.deleteTempDir
 import io.realm.test.platform.PlatformUtils.triggerGC
@@ -36,6 +37,8 @@ import kotlin.test.Ignore
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
+// New memory model doesn't seem to fix our memory leak, or maybe we have to reevaluate how we do
+// the tests.
 class MemoryTests {
 
     lateinit var tmpDir: String
@@ -52,8 +55,11 @@ class MemoryTests {
 
     // TODO Only run on macOS, filter using https://developer.apple.com/documentation/foundation/nsprocessinfo/3608556-iosapponmac when upgrading to XCode 12
     @Test
-    @Ignore // We currently do not clean up intermediate versions if the realm itself is garbage collected
+    @Ignore // Investigate https://github.com/realm/realm-kotlin/issues/327
     fun garbageCollectorShouldFreeNativeResources() {
+        @OptIn(ExperimentalStdlibApi::class)
+        println("NEW_MEMORY_MODEL: " + isExperimentalMM())
+
         val referenceHolder = mutableListOf<Sample>()
         val amountOfMemoryMappedInProcessCMD =
             "vmmap  -summary ${platform.posix.getpid()}  2>/dev/null | awk '/mapped/ {print \$3}'";
@@ -61,7 +67,9 @@ class MemoryTests {
             val realm = openRealmFromTmpDir()
             // TODO use Realm.delete once this is implemented
             realm.writeBlocking {
-                objects(Sample::class).delete()
+                query<Sample>()
+                    .find()
+                    .delete()
             }
 
             // allocating a 1 MB string
@@ -109,6 +117,9 @@ class MemoryTests {
     // TODO Only run on macOS, filter using https://developer.apple.com/documentation/foundation/nsprocessinfo/3608556-iosapponmac when upgrading to XCode 12
     @Test
     fun closeShouldFreeMemory() {
+        @OptIn(ExperimentalStdlibApi::class)
+        println("NEW_MEMORY_MODEL: " + isExperimentalMM())
+
         val referenceHolder = mutableListOf<Sample>()
         val amountOfMemoryMappedInProcessCMD =
             "vmmap  -summary ${platform.posix.getpid()}  2>/dev/null | awk '/mapped/ {print \$3}'";